update screening structure
